chore: clean up PlayerAvatar code
