+++ game/src/player/mod.rs
@@ -1,4 +1,4 @@
-mod entity;
+mod avatar;
 mod menu_control;
 mod movement;
 mod player_control;
@@ -6,5 +6,5 @@ mod player_creation;
 mod player_sprite;
 mod plugin;
 
-pub use entity::Player;
+pub use avatar::Player;
 pub use plugin::PlayerPlugin;
